# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CyberShield 网络攻防平台 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: List security alerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved, all]
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of security alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a security alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "201":
          description: Created alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    post:
      operationId: resolveAlert
      tags: [alerts]
      summary: Resolve a security alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Alert not found
...
  /connections:
    get:
      operationId: getConnections
      tags: [connections]
      summary: List monitored network connections
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, blocked, suspicious, all]
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NetworkConnection"
  /connections/{id}/block:
    post:
      operationId: blockConnection
      tags: [connections]
      summary: Block a network connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blocked connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkConnection"
        "404":
          description: Connection not found
...
  /threat-intel/indicators:
    get:
      operationId: getThreatIndicators
      tags: [threatIntel]
      summary: List threat indicators
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [ip, domain, hash, all]
      responses:
        "200":
          description: List of threat indicators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreatIndicator"
    post:
      operationId: createThreatIndicator
      tags: [threatIntel]
      summary: Create a threat indicator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThreatIndicatorInput"
      responses:
        "201":
          description: Created indicator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatIndicator"
  /threat-intel/indicators/{id}:
    patch:
      operationId: updateThreatIndicator
      tags: [threatIntel]
      summary: Update a threat indicator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThreatIndicatorUpdate"
      responses:
        "200":
          description: Updated indicator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatIndicator"
        "404":
          description: Indicator not found
    delete:
      operationId: deleteThreatIndicator
      tags: [threatIntel]
      summary: Delete a threat indicator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Indicator not found
...
  /threat-intel/indicators/{id}/block:
    post:
      operationId: blockThreatIndicator
      tags: [threatIntel]
      summary: Push an IP indicator to the firewall as a deny rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Created firewall rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirewallRule"
        "400":
          description: Indicator is not an IP type
        "404":
          description: Indicator not found
        "409":
          description: Firewall rule already exists for this IP
components:
  schemas:
...
    Alert:
      type: object
      required:
        - id
        - title
        - description
        - severity
        - type
        - sourceIp
        - status
        - detectedAt
        - resolvedAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        type:
          type: string
          enum: [intrusion, ddos, malware, brute_force, port_scan, sql_injection, xss, other]
        sourceIp:
          type: string
        targetPort:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, resolved]
        detectedAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
        knownThreat:
          type: boolean
          description: Source IP matches a known threat indicator
    AlertInput:
      type: object
      required:
        - title
        - description
        - severity
        - type
        - sourceIp
      properties:
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        type:
          type: string
          enum: [intrusion, ddos, malware, brute_force, port_scan, sql_injection, xss, other]
        sourceIp:
          type: string
        targetPort:
          type: ["integer", "null"]
    FirewallRule:
      type: object
      required:
        - id
        - name
        - action
        - protocol
        - sourceIp
        - destinationPort
        - enabled
        - priority
        - createdAt
        - description
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        action:
          type: string
          enum: [allow, deny, drop, log]
        protocol:
          type: string
          enum: [tcp, udp, icmp, any]
        sourceIp:
          type: string
          description: CIDR notation or specific IP, * for any
        destinationPort:
          type: string
          description: Port number, range, or * for any
        enabled:
          type: boolean
        priority:
          type: integer
        createdAt:
          type: string
...
    NetworkConnection:
      type: object
      required:
        - id
        - sourceIp
        - sourcePort
        - destinationIp
        - destinationPort
        - protocol
        - status
        - bytesIn
        - bytesOut
        - country
        - connectedAt
      properties:
        id:
          type: integer
        sourceIp:
          type: string
        sourcePort:
          type: integer
        destinationIp:
          type: string
        destinationPort:
          type: integer
        protocol:
          type: string
          enum: [tcp, udp, icmp]
        status:
          type: string
          enum: [active, blocked, suspicious]
        bytesIn:
          type: integer
        bytesOut:
          type: integer
        country:
          type: string
        connectedAt:
          type: string
        knownThreat:
          type: boolean
          description: Source IP matches a known threat indicator
    ThreatIndicator:
      type: object
      required:
        - id
        - type
        - value
        - category
        - confidence
        - source
        - description
        - firstSeen
        - lastSeen
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [ip, domain, hash]
        value:
          type: string
        category:
          type: string
          enum: [malware, botnet, phishing, c2, scanner, spam, apt, other]
        confidence:
          type: integer
          description: Confidence 0-100
        source:
          type: string
        description:
          type: string
        firstSeen:
          type: string
        lastSeen:
          type: string
    ThreatIndicatorInput:
      type: object
      required:
        - type
        - value
        - category
      properties:
        type:
          type: string
          enum: [ip, domain, hash]
        value:
          type: string
          minLength: 1
        category:
          type: string
          enum: [malware, botnet, phishing, c2, scanner, spam, apt, other]
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        source:
          type: string
        description:
          type: string
    ThreatIndicatorUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [ip, domain, hash]
        value:
          type: string
          minLength: 1
        category:
          type: string
          enum: [malware, botnet, phishing, c2, scanner, spam, apt, other]
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        source:
          type: string
        description:
          type: string
...