# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CyberShield 网络攻防平台 API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from './alertSeverity';
import type { AlertStatus } from './alertStatus';
import type { AlertType } from './alertType';

export interface Alert {
  id: number;
  title: string;
  description: string;
  severity: AlertSeverity;
  type: AlertType;
  sourceIp: string;
  /** @nullable */
  targetPort?: number | null;
  status: AlertStatus;
  detectedAt: string;
  /** @nullable */
  resolvedAt: string | null;
  /** Source IP matches a known threat indicator */
  knownThreat?: boolean;
}