# lib/api-zod/src/generated/types/networkConnection.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CyberShield 网络攻防平台 API
 * OpenAPI spec version: 0.1.0
 */
import type { NetworkConnectionProtocol } from './networkConnectionProtocol';
import type { NetworkConnectionStatus } from './networkConnectionStatus';

export interface NetworkConnection {
  id: number;
  sourceIp: string;
  sourcePort: number;
  destinationIp: string;
  destinationPort: number;
  protocol: NetworkConnectionProtocol;
  status: NetworkConnectionStatus;
  bytesIn: number;
  bytesOut: number;
  country: string;
  connectedAt: string;
  /** Source IP matches a known threat indicator */
  knownThreat?: boolean;
}