# lib/api-zod/src/generated/types/threatIndicator.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CyberShield 网络攻防平台 API
 * OpenAPI spec version: 0.1.0
 */
import type { ThreatIndicatorCategory } from './threatIndicatorCategory';
import type { ThreatIndicatorType } from './threatIndicatorType';

export interface ThreatIndicator {
  id: number;
  type: ThreatIndicatorType;
  value: string;
  category: ThreatIndicatorCategory;
  /** Confidence 0-100 */
  confidence: number;
  source: string;
  description: string;
  firstSeen: string;
  lastSeen: string;
}