# lib/api-zod/src/generated/types/threatIndicatorInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CyberShield 网络攻防平台 API
 * OpenAPI spec version: 0.1.0
 */
import type { ThreatIndicatorInputCategory } from './threatIndicatorInputCategory';
import type { ThreatIndicatorInputType } from './threatIndicatorInputType';

export interface ThreatIndicatorInput {
  type: ThreatIndicatorInputType;
  /** @minLength 1 */
  value: string;
  category: ThreatIndicatorInputCategory;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence?: number;
  source?: string;
  description?: string;
}