# lib/api-zod/src/generated/types/threatIndicatorUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CyberShield 网络攻防平台 API
 * OpenAPI spec version: 0.1.0
 */
import type { ThreatIndicatorUpdateCategory } from './threatIndicatorUpdateCategory';
import type { ThreatIndicatorUpdateType } from './threatIndicatorUpdateType';

export interface ThreatIndicatorUpdate {
  type?: ThreatIndicatorUpdateType;
  /** @minLength 1 */
  value?: string;
  category?: ThreatIndicatorUpdateCategory;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence?: number;
  source?: string;
  description?: string;
}